feat: move renderring of h1 and textarea to separate function to declare textarea in consructor, create multiplyRows method for easier renderring the keyboard
